fix: fixed export of modules and help structure of files
