--- .release-it.json
@@ -1,8 +1,0 @@
-{
-  "git": {
-    "commitMessage": "chore(): release v${version}"
-  },
-  "github": {
-    "release": true
-  }
-}
