+++ lib/modules/sms/contract/index.ts
@@ -1,0 +1,1 @@
+export type { SendSMSInput } from './types';
