+++ lib/modules/sms/contract/types/index.ts
@@ -1,0 +1,1 @@
+export type { SendSMSInput } from './sms.types';
