+++ lib/modules/sms/contract/types/sms.types.ts
@@ -1,0 +1,5 @@
+export type SendSMSInput = {
+  Message: string;
+  PhoneNumber: string;
+  Subject?: string;
+};
