+++ lib/modules/sms/index.ts
@@ -1,0 +1,2 @@
+export type { SendSMSInput } from './contract';
+export * from './services';
