+++ lib/modules/sms/services/index.ts
@@ -1,0 +1,1 @@
+export * from './sms.service';
