+++ lib/modules/sns/common/constants/sns.constants.ts
