+++ lib/modules/sns/common/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './fixtures';
