+++ lib/modules/sns/contract/index.ts
@@ -1,0 +1,9 @@
+export { SnsOptionsFactory, SnsOptions, SnsAsyncOptions } from './interfaces';
+export type {
+  CreateTopicResponse,
+  PublishResponse,
+  SubscribeInput,
+  SubscribeCommandOutput,
+  PublishInput,
+  CreateTopicInput,
+} from './types';
