+++ lib/modules/sns/contract/interfaces/index.ts
@@ -1,0 +1,1 @@
+export { SnsOptionsFactory, SnsOptions, SnsAsyncOptions } from './sns-module.interfaces';
