+++ lib/modules/sns/contract/interfaces/sns-module.interfaces.ts
