+++ lib/modules/sns/contract/types/index.ts
@@ -1,0 +1,8 @@
+export type {
+  CreateTopicResponse,
+  PublishResponse,
+  SubscribeInput,
+  SubscribeCommandOutput,
+  PublishInput,
+  CreateTopicInput,
+} from './sns.types';
