+++ lib/modules/sns/contract/types/sns.types.ts
@@ -8,9 +8,3 @@ export type SubscribeCommandOutput = sns.SubscribeCommandOutput;
 
 export type PublishInput = sns.PublishInput;
 export type CreateTopicInput = sns.CreateTopicInput;
-
-export type SendSMSInput = {
-  Message: string;
-  PhoneNumber: string;
-  Subject?: string;
-};
