+++ lib/modules/sns/index.ts
@@ -1,0 +1,14 @@
+export * from './common/constants';
+export * from './common/testing';
+
+export * from './services';
+
+export { SnsOptionsFactory, SnsOptions, SnsAsyncOptions } from './contract/interfaces';
+export type {
+  CreateTopicResponse,
+  PublishResponse,
+  SubscribeInput,
+  SubscribeCommandOutput,
+  PublishInput,
+  CreateTopicInput,
+} from './contract/types';
