+++ lib/modules/sns/sns.module.ts
@@ -1,0 +1,33 @@
+import { DynamicModule, Module } from '@nestjs/common';
+
+import { SnsService } from './services';
+import { SmsService } from '../sms/services/sms.service';
+
+import { SNS_OPTIONS } from './common/constants';
+
+import type { SnsOptions } from './contract';
+
+@Module({
+  imports: [],
+  providers: [SnsService, SmsService],
+  exports: [SnsService, SmsService],
+})
+export class SnsModule {
+  static register(options: SnsOptions): DynamicModule {
+    const { isGlobal, ...snsOptions } = options;
+    isGlobal;
+    return {
+      module: SnsModule,
+      providers: [
+        {
+          provide: SNS_OPTIONS,
+          useValue: snsOptions,
+        },
+        SnsService,
+        SmsService,
+      ],
+      exports: [SnsService, SmsService],
+      global: isGlobal,
+    };
+  }
+}
